# Melhoria do sistema Cache - Performance # Melhoria do código em algumas classes # adicionado no formato do money # Fix do comando TPA nas versões a abaixo da 1.7.10 # Rework dos inventários do plugin (interno) e adição no menu do playtime o seu tempo jogado # Fix do playtime mostrar números infinitos

+++ src/main/kotlin/github/gilbertokpl/total/commands/CommandMoney.kt
@@ -56,7 +56,7 @@ class CommandMoney : github.gilbertokpl.core.external.command.CommandCreator("mo
             for (i in MoneyUtil.tycoonPlayer) {
                 s.sendMessage(
                     MoneyUtil.coinReplacer(
-                        LangConfig.moneyTop, i.value.toDouble()
+                        LangConfig.moneyTop, i.value
                     )
                         .replace("%player%", i.key)
                         .replace("%position%", position.toString())
